+++ include/functions.h
@@ -148,7 +148,7 @@ void OtherFX_Stop1(int soundID_count);
 void OtherFX_Stop2(int soundID_count);
 void OtherFX_RecycleNew(u32 *soundID_Count, u32 newSoundID, u32 modifyFlags);
 void OtherFX_RecycleMute(u32 *soundID_Count);
-b32 EngineAudio_InitOnce(u32 soundID, u32 flags);
+b32 EngineAudio_InitOnce(u16 soundID, u32 flags);
 s16 EngineAudio_Recalculate(u32 soundID, u32 sfx);
 void EngineAudio_Stop(u32 soundID);
 void SetReverbMode(u16 newReverbMode);
